all socs: DPL: Fix MISRA C 2012 Rule 10.1 issue
MISRA.ETYPE.INAPPR.OPERAND.UNOP.2012

Operands shall not be of an inappropriate essential type

Fixes: PDK-18979

+++ source/kernel/nortos/dpl/c75/HwiP_c75.c
@@ -474,7 +474,7 @@ char *Hwi_getIsrStackAddress(void)
     isrStack += (uint64_t)_symval(&__TI_STACK_SIZE);
     isrStack -= 0x1;
 
-    isrStack &= ~0x7;   /* align to long word */
+    isrStack &= ~((uint64_t)(0x7));   /* align to long word */
 
     return ((char *)isrStack);
 }
